refactor(rsdoctor): inline rslib runtime and disable runtimeChunk
